Move enlisting to utils module
